CreateToImplementLinkedFileAllocation.c
Linked file allocation is a file allocation method used in operating systems to store files on disk. In this approach, a file is represented as a linked list of disk blocks. Each block contains a pointer to the next block in the sequence, which ensures that the file blocks do not need to be contiguous on disk.